+++ _updateHeading.js
@@ -1,3 +1,4 @@
+import isFunction from './utils/types/isFunction'
 import trim from './utils/lang/trim'
 import createElement from './utils/dom/createElement'
 import setAttributes from './utils/dom/setAttributes'
@@ -11,6 +12,7 @@ const _updateHeading = ($heading, i, options) => {
   const showCode = options.showCode || false
   const chapterCode = options.chapterCode || ''
   const anchorURL = options.anchorURL || ''
+  const anchorLinkFilter = options.anchorLinkFilter || null
   const headingId = `heading-${i}`
   const attrs = {
     id: headingId,
@@ -18,6 +20,7 @@ const _updateHeading = ($heading, i, options) => {
     'data-id': i
   }
   const text = trim($heading.innerHTML)
+  let url
   let $anchor
   let $icon
 
@@ -31,13 +34,23 @@ const _updateHeading = ($heading, i, options) => {
     return false
   }
 
+  if (anchorURL) {
+    url = anchorURL
+  } else {
+    if (isFunction(anchorLinkFilter)) {
+      url = anchorLinkFilter($heading.tagName.toLowerCase(), text, i)
+    } else {
+      url = `#${headingId}`
+    }
+  }
+
   $icon = icon('hash', { iconSet: 'outline' })
   $anchor = createElement(
     'a',
     {
       id: `anchor-${i}`,
       className: `${CLS_HEADING}__anchor anchor-${i}`,
-      href: anchorURL || `#${headingId}`,
+      href: url,
       target: anchorURL ? '_blank' : 'self',
       'data-id': i
     },